Add integer type support for aten.max.int op.

+++ python/torch_mlir_e2e_test/test_suite/reduction.py
@@ -303,7 +303,7 @@ class ReduceMaxAlongDim(torch.nn.Module):
     @export
     @annotate_args([
         None,
-        ([-1, -1, -1], torch.float64, True),
+        ([-1, -1, -1], torch.int64, True),
     ])
     def forward(self, a):
         return torch.ops.aten.max(a, 1)[0]
@@ -311,7 +311,7 @@ class ReduceMaxAlongDim(torch.nn.Module):
 
 @register_test_case(module_factory=lambda: ReduceMaxAlongDim())
 def ReduceMaxAlongDim_basic(module, tu: TestUtils):
-    module.forward(tu.rand(3, 4, 5).to(torch.float64))
+    module.forward(tu.rand(3, 4, 5).to(torch.int64))
 
 # ==============================================================================
